+++ packages/jest-circus/src/__mocks__/testUtils.ts
@@ -5,10 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as fs from 'fs';
 import {tmpdir} from 'os';
 import * as path from 'path';
 import {createHash} from 'crypto';
+import * as fs from 'graceful-fs';
 // eslint-disable-next-line import/named
 import {ExecaSyncReturnValue, sync as spawnSync} from 'execa';
 import {skipSuiteOnWindows} from '@jest/test-utils';