+++ packages/jest-haste-map/src/types.ts
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import type {Stats} from 'fs';
+import type {Stats} from 'graceful-fs';
 import type {Config} from '@jest/types';
 import type ModuleMap from './ModuleMap';
 import type HasteFS from './HasteFS';