+++ packages/jest-runtime/src/index.ts
@@ -5,9 +5,19 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {URL, fileURLToPath} from 'url';
+import {URL, fileURLToPath, pathToFileURL} from 'url';
 import * as path from 'path';
-import {Script, compileFunction} from 'vm';
+import {
+  Script,
+  // @ts-ignore: experimental, not added to the types
+  SourceTextModule,
+  // @ts-ignore: experimental, not added to the types
+  SyntheticModule,
+  Context as VMContext,
+  // @ts-ignore: experimental, not added to the types
+  Module as VMModule,
+  compileFunction,
+} from 'vm';
 import * as nativeModule from 'module';
 import type {Config, Global} from '@jest/types';
 import type {
@@ -44,8 +54,8 @@ import Snapshot = require('jest-snapshot');
 import stripBOM = require('strip-bom');
 
 interface JestGlobalsValues extends Global.TestFrameworkGlobals {
-  jest: JestGlobals.jest;
-  expect: JestGlobals.expect;
+  jest: typeof JestGlobals.jest;
+  expect: typeof JestGlobals.expect;
 }
 
 type HasteMapOptions = {
@@ -73,8 +83,17 @@ type InitialModule = Partial<Module> &
 type ModuleRegistry = Map<string, InitialModule | Module>;
 type ResolveOptions = Parameters<typeof require.resolve>[1];
 
-type BooleanObject = Record<string, boolean>;
-type CacheFS = {[path: string]: string};
+type StringMap = Map<string, string>;
+type BooleanMap = Map<string, boolean>;
+
+const fromEntries: typeof Object.fromEntries =
+  Object.fromEntries ??
+  function fromEntries<T>(iterable: Iterable<[string, T]>) {
+    return [...iterable].reduce<Record<string, T>>((obj, [key, val]) => {
+      obj[key] = val;
+      return obj;
+    }, {});
+  };
 
 namespace Runtime {
   export type Context = JestContext;
@@ -106,38 +125,41 @@ const EVAL_RESULT_VARIABLE = 'Object.<anonymous>';
 
 type RunScriptEvalResult = {[EVAL_RESULT_VARIABLE]: ModuleWrapper};
 
+const runtimeSupportsVmModules = typeof SyntheticModule === 'function';
+
 /* eslint-disable-next-line no-redeclare */
 class Runtime {
-  private _cacheFS: CacheFS;
+  private _cacheFS: StringMap;
   private _config: Config.ProjectConfig;
   private _coverageOptions: ShouldInstrumentOptions;
   private _currentlyExecutingModulePath: string;
   private _environment: JestEnvironment;
-  private _explicitShouldMock: BooleanObject;
+  private _explicitShouldMock: BooleanMap;
   private _internalModuleRegistry: ModuleRegistry;
   private _isCurrentlyExecutingManualMock: string | null;
-  private _mockFactories: Record<string, () => unknown>;
-  private _mockMetaDataCache: {
-    [key: string]: jestMock.MockFunctionMetadata<unknown, Array<unknown>>;
-  };
+  private _mockFactories: Map<string, () => unknown>;
+  private _mockMetaDataCache: Map<
+    string,
+    jestMock.MockFunctionMetadata<unknown, Array<unknown>>
+  >;
   private _mockRegistry: Map<string, any>;
   private _isolatedMockRegistry: Map<string, any> | null;
   private _moduleMocker: typeof jestMock;
   private _isolatedModuleRegistry: ModuleRegistry | null;
   private _moduleRegistry: ModuleRegistry;
-  private _needsCoverageMapped: Set<string>;
+  private _esmoduleRegistry: Map<string, Promise<VMModule>>;
   private _resolver: Resolver;
   private _shouldAutoMock: boolean;
-  private _shouldMockModuleCache: BooleanObject;
-  private _shouldUnmockTransitiveDependenciesCache: BooleanObject;
-  private _sourceMapRegistry: SourceMapRegistry;
+  private _shouldMockModuleCache: BooleanMap;
+  private _shouldUnmockTransitiveDependenciesCache: BooleanMap;
+  private _sourceMapRegistry: StringMap;
   private _scriptTransformer: ScriptTransformer;
   private _fileTransforms: Map<string, TransformResult>;
   private _v8CoverageInstrumenter: CoverageInstrumenter | undefined;
   private _v8CoverageResult: V8Coverage | undefined;
-  private _transitiveShouldMock: BooleanObject;
+  private _transitiveShouldMock: BooleanMap;
   private _unmockList: RegExp | undefined;
-  private _virtualMocks: BooleanObject;
+  private _virtualMocks: BooleanMap;
   private _moduleImplementation?: typeof nativeModule.Module;
   private jestObjectCaches: Map<string, Jest>;
 
@@ -145,10 +167,10 @@ class Runtime {
     config: Config.ProjectConfig,
     environment: JestEnvironment,
     resolver: Resolver,
-    cacheFS?: CacheFS,
+    cacheFS: Record<string, string> = {},
     coverageOptions?: ShouldInstrumentOptions,
   ) {
-    this._cacheFS = cacheFS || Object.create(null);
+    this._cacheFS = new Map(Object.entries(cacheFS));
     this._config = config;
     this._coverageOptions = coverageOptions || {
       changedFiles: undefined,
@@ -156,32 +178,33 @@ class Runtime {
       collectCoverageFrom: [],
       collectCoverageOnlyFrom: undefined,
       coverageProvider: 'babel',
+      sourcesRelatedToTestsInChangedFiles: undefined,
     };
     this._currentlyExecutingModulePath = '';
     this._environment = environment;
-    this._explicitShouldMock = Object.create(null);
+    this._explicitShouldMock = new Map();
     this._internalModuleRegistry = new Map();
     this._isCurrentlyExecutingManualMock = null;
-    this._mockFactories = Object.create(null);
+    this._mockFactories = new Map();
     this._mockRegistry = new Map();
     // during setup, this cannot be null (and it's fine to explode if it is)
     this._moduleMocker = this._environment.moduleMocker!;
     this._isolatedModuleRegistry = null;
     this._isolatedMockRegistry = null;
     this._moduleRegistry = new Map();
-    this._needsCoverageMapped = new Set();
+    this._esmoduleRegistry = new Map();
     this._resolver = resolver;
     this._scriptTransformer = new ScriptTransformer(config);
     this._shouldAutoMock = config.automock;
-    this._sourceMapRegistry = Object.create(null);
+    this._sourceMapRegistry = new Map();
     this._fileTransforms = new Map();
-    this._virtualMocks = Object.create(null);
+    this._virtualMocks = new Map();
     this.jestObjectCaches = new Map();
 
-    this._mockMetaDataCache = Object.create(null);
-    this._shouldMockModuleCache = Object.create(null);
-    this._shouldUnmockTransitiveDependenciesCache = Object.create(null);
-    this._transitiveShouldMock = Object.create(null);
+    this._mockMetaDataCache = new Map();
+    this._shouldMockModuleCache = new Map();
+    this._shouldUnmockTransitiveDependenciesCache = new Map();
+    this._transitiveShouldMock = new Map();
 
     this._unmockList = unmockRegExpCache.get(config);
     if (!this._unmockList && config.unmockedModulePathPatterns) {
@@ -192,13 +215,11 @@ class Runtime {
     }
 
     if (config.automock) {
+      const virtualMocks = fromEntries(this._virtualMocks);
       config.setupFiles.forEach(filePath => {
         if (filePath && filePath.includes(NODE_MODULES)) {
-          const moduleID = this._resolver.getModuleID(
-            this._virtualMocks,
-            filePath,
-          );
-          this._transitiveShouldMock[moduleID] = false;
+          const moduleID = this._resolver.getModuleID(virtualMocks, filePath);
+          this._transitiveShouldMock.set(moduleID, false);
         }
       });
     }
@@ -281,7 +302,6 @@ class Runtime {
     moduleMap: HasteMap.ModuleMap,
   ): Resolver {
     return new Resolver(moduleMap, {
-      browser: config.browser,
       defaultPlatform: config.haste.defaultPlatform,
       extensions: config.moduleFileExtensions.map(extension => '.' + extension),
       hasCoreModules: true,
@@ -302,6 +322,141 @@ class Runtime {
     return cliOptions;
   }
 
+  // unstable as it should be replaced by https://github.com/nodejs/modules/issues/393, and we don't want people to use it
+  unstable_shouldLoadAsEsm = Resolver.unstable_shouldLoadAsEsm;
+
+  private async loadEsmModule(
+    modulePath: Config.Path,
+    query = '',
+  ): Promise<VMModule> {
+    const cacheKey = modulePath + query;
+
+    if (!this._esmoduleRegistry.has(cacheKey)) {
+      invariant(
+        typeof this._environment.getVmContext === 'function',
+        'ES Modules are only supported if your test environment has the `getVmContext` function',
+      );
+
+      const context = this._environment.getVmContext();
+
+      invariant(context);
+
+      if (this._resolver.isCoreModule(modulePath)) {
+        const core = await this._importCoreModule(modulePath, context);
+        this._esmoduleRegistry.set(cacheKey, core);
+        return core;
+      }
+
+      const transformedCode = this.transformFile(modulePath, {
+        isInternalModule: false,
+        supportsDynamicImport: true,
+        supportsStaticESM: true,
+      });
+
+      const module = new SourceTextModule(transformedCode, {
+        context,
+        identifier: modulePath,
+        importModuleDynamically: this.linkModules.bind(this),
+        initializeImportMeta(meta: ImportMeta) {
+          meta.url = pathToFileURL(modulePath).href;
+        },
+      });
+
+      this._esmoduleRegistry.set(
+        cacheKey,
+        // we wanna put the linking promise in the cache so modules loaded in
+        // parallel can all await it. We then await it synchronously below, so
+        // we shouldn't get any unhandled rejections
+        module
+          .link(this.linkModules.bind(this))
+          .then(() => module.evaluate())
+          .then(() => module),
+      );
+    }
+
+    const module = this._esmoduleRegistry.get(cacheKey);
+
+    invariant(module);
+
+    return module;
+  }
+
+  private linkModules(specifier: string, referencingModule: VMModule) {
+    if (specifier === '@jest/globals') {
+      const fromCache = this._esmoduleRegistry.get('@jest/globals');
+
+      if (fromCache) {
+        return fromCache;
+      }
+      const globals = this.getGlobalsForEsm(
+        referencingModule.identifier,
+        referencingModule.context,
+      );
+      this._esmoduleRegistry.set('@jest/globals', globals);
+
+      return globals;
+    }
+
+    const [path, query] = specifier.split('?');
+
+    const resolved = this._resolveModule(referencingModule.identifier, path);
+
+    if (
+      this._resolver.isCoreModule(resolved) ||
+      this.unstable_shouldLoadAsEsm(resolved)
+    ) {
+      return this.loadEsmModule(resolved, query);
+    }
+
+    return this.loadCjsAsEsm(
+      referencingModule.identifier,
+      resolved,
+      referencingModule.context,
+    );
+  }
+
+  async unstable_importModule(
+    from: Config.Path,
+    moduleName?: string,
+  ): Promise<void> {
+    invariant(
+      runtimeSupportsVmModules,
+      'You need to run with a version of node that supports ES Modules in the VM API.',
+    );
+
+    const [path, query] = (moduleName ?? '').split('?');
+
+    const modulePath = this._resolveModule(from, path);
+
+    return this.loadEsmModule(modulePath, query);
+  }
+
+  private async loadCjsAsEsm(
+    from: Config.Path,
+    modulePath: Config.Path,
+    context: VMContext,
+  ) {
+    // CJS loaded via `import` should share cache with other CJS: https://github.com/nodejs/modules/issues/503
+    const cjs = this.requireModuleOrMock(from, modulePath);
+
+    const module = new SyntheticModule(
+      ['default'],
+      function () {
+        // @ts-ignore: TS doesn't know what `this` is
+        this.setExport('default', cjs);
+      },
+      {context, identifier: modulePath},
+    );
+
+    await module.link(() => {
+      throw new Error('This should never happen');
+    });
+
+    await module.evaluate();
+
+    return module;
+  }
+
   requireModule<T = unknown>(
     from: Config.Path,
     moduleName?: string,
@@ -309,7 +464,7 @@ class Runtime {
     isRequireActual?: boolean | null,
   ): T {
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
@@ -321,19 +476,14 @@ class Runtime {
     const manualMock =
       moduleName && this._resolver.getMockModule(from, moduleName);
     if (
-      (!options || !options.isInternalModule) &&
+      !options?.isInternalModule &&
       !isRequireActual &&
       !moduleResource &&
       manualMock &&
       manualMock !== this._isCurrentlyExecutingManualMock &&
-      this._explicitShouldMock[moduleID] !== false
+      this._explicitShouldMock.get(moduleID) !== false
     ) {
       modulePath = manualMock;
-    }
-
-    if (moduleName === '@jest/globals') {
-      // @ts-ignore: we don't care that it's not assignable to T
-      return this.getGlobalsForFile(from);
     }
 
     if (moduleName && this._resolver.isCoreModule(moduleName)) {
@@ -346,7 +496,9 @@ class Runtime {
 
     let moduleRegistry;
 
-    if (!options || !options.isInternalModule) {
+    if (options?.isInternalModule) {
+      moduleRegistry = this._internalModuleRegistry;
+    } else {
       if (
         this._moduleRegistry.get(modulePath) ||
         !this._isolatedModuleRegistry
@@ -355,8 +507,6 @@ class Runtime {
       } else {
         moduleRegistry = this._isolatedModuleRegistry;
       }
-    } else {
-      moduleRegistry = this._internalModuleRegistry;
     }
 
     const module = moduleRegistry.get(modulePath);
@@ -402,7 +552,7 @@ class Runtime {
 
   requireMock<T = unknown>(from: Config.Path, moduleName: string): T {
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
@@ -418,8 +568,9 @@ class Runtime {
 
     const mockRegistry = this._isolatedMockRegistry || this._mockRegistry;
 
-    if (moduleID in this._mockFactories) {
-      const module = this._mockFactories[moduleID]();
+    if (this._mockFactories.has(moduleID)) {
+      // has check above makes this ok
+      const module = this._mockFactories.get(moduleID)!();
       mockRegistry.set(moduleID, module);
       return module as T;
     }
@@ -496,7 +647,7 @@ class Runtime {
     moduleRegistry: ModuleRegistry,
   ) {
     if (path.extname(modulePath) === '.json') {
-      const text = stripBOM(fs.readFileSync(modulePath, 'utf8'));
+      const text = stripBOM(this.readFile(modulePath));
 
       const transformedFile = this._scriptTransformer.transformJson(
         modulePath,
@@ -526,12 +677,18 @@ class Runtime {
     };
   }
 
-  requireModuleOrMock(from: Config.Path, moduleName: string): unknown {
+  requireModuleOrMock<T = unknown>(from: Config.Path, moduleName: string): T {
+    // this module is unmockable
+    if (moduleName === '@jest/globals') {
+      // @ts-ignore: we don't care that it's not assignable to T
+      return this.getGlobalsForCjs(from);
+    }
+
     try {
       if (this._shouldMock(from, moduleName)) {
-        return this.requireMock(from, moduleName);
+        return this.requireMock<T>(from, moduleName);
       } else {
-        return this.requireModule(from, moduleName);
+        return this.requireModule<T>(from, moduleName);
       }
     } catch (e) {
       const moduleNotFound = Resolver.tryCastModuleNotFoundError(e);
@@ -567,16 +724,22 @@ class Runtime {
     try {
       fn();
     } finally {
+      // might be cleared within the callback
+      this._isolatedModuleRegistry?.clear();
+      this._isolatedMockRegistry?.clear();
       this._isolatedModuleRegistry = null;
       this._isolatedMockRegistry = null;
     }
   }
 
   resetModules(): void {
+    this._isolatedModuleRegistry?.clear();
+    this._isolatedMockRegistry?.clear();
     this._isolatedModuleRegistry = null;
     this._isolatedMockRegistry = null;
     this._mockRegistry.clear();
     this._moduleRegistry.clear();
+    this._esmoduleRegistry.clear();
 
     if (this._environment) {
       if (this._environment.global) {
@@ -641,24 +804,13 @@ class Runtime {
       });
   }
 
-  getSourceMapInfo(coveredFiles: Set<string>): Record<string, string> {
-    return Object.keys(this._sourceMapRegistry).reduce<Record<string, string>>(
-      (result, sourcePath) => {
-        if (
-          coveredFiles.has(sourcePath) &&
-          this._needsCoverageMapped.has(sourcePath) &&
-          fs.existsSync(this._sourceMapRegistry[sourcePath])
-        ) {
-          result[sourcePath] = this._sourceMapRegistry[sourcePath];
-        }
-        return result;
-      },
-      {},
-    );
+  // TODO - remove in Jest 26
+  getSourceMapInfo(_coveredFiles: Set<string>): Record<string, string> {
+    return {};
   }
 
   getSourceMaps(): SourceMapRegistry {
-    return this._sourceMapRegistry;
+    return fromEntries(this._sourceMapRegistry);
   }
 
   setMock(
@@ -667,17 +819,18 @@ class Runtime {
     mockFactory: () => unknown,
     options?: {virtual?: boolean},
   ): void {
-    if (options && options.virtual) {
+    if (options?.virtual) {
       const mockPath = this._resolver.getModulePath(from, moduleName);
-      this._virtualMocks[mockPath] = true;
+
+      this._virtualMocks.set(mockPath, true);
     }
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
-    this._explicitShouldMock[moduleID] = true;
-    this._mockFactories[moduleID] = mockFactory;
+    this._explicitShouldMock.set(moduleID, true);
+    this._mockFactories.set(moduleID, mockFactory);
   }
 
   restoreAllMocks(): void {
@@ -690,6 +843,32 @@ class Runtime {
 
   clearAllMocks(): void {
     this._moduleMocker.clearAllMocks();
+  }
+
+  teardown(): void {
+    this.restoreAllMocks();
+    this.resetAllMocks();
+    this.resetModules();
+
+    this._internalModuleRegistry.clear();
+    this._mockFactories.clear();
+    this._mockMetaDataCache.clear();
+    this._shouldMockModuleCache.clear();
+    this._shouldUnmockTransitiveDependenciesCache.clear();
+    this._explicitShouldMock.clear();
+    this._transitiveShouldMock.clear();
+    this._virtualMocks.clear();
+    this._cacheFS.clear();
+    this._unmockList = undefined;
+
+    this._sourceMapRegistry.clear();
+
+    this._fileTransforms.clear();
+    this.jestObjectCaches.clear();
+
+    this._v8CoverageResult = [];
+    this._v8CoverageInstrumenter = undefined;
+    this._moduleImplementation = undefined;
   }
 
   private _resolveModule(from: Config.Path, to?: string) {
@@ -795,23 +974,8 @@ class Runtime {
     Object.defineProperty(localModule, 'require', {
       value: this._createRequireImplementation(localModule, options),
     });
-    const transformedFile = this._scriptTransformer.transform(
-      filename,
-      this._getFullTransformationOptions(options),
-      this._cacheFS[filename],
-    );
 
-    // we only care about non-internal modules
-    if (!options || !options.isInternalModule) {
-      this._fileTransforms.set(filename, transformedFile);
-    }
-
-    if (transformedFile.sourceMapPath) {
-      this._sourceMapRegistry[filename] = transformedFile.sourceMapPath;
-      if (transformedFile.mapCoverage) {
-        this._needsCoverageMapped.add(filename);
-      }
-    }
+    const transformedCode = this.transformFile(filename, options);
 
     let compiledFunction: ModuleWrapper | null = null;
 
@@ -823,7 +987,7 @@ class Runtime {
         if (typeof compileFunction === 'function') {
           try {
             compiledFunction = compileFunction(
-              transformedFile.code,
+              transformedCode,
               this.constructInjectedModuleParameters(),
               {
                 filename,
@@ -834,10 +998,7 @@ class Runtime {
             throw handlePotentialSyntaxError(e);
           }
         } else {
-          const script = this.createScriptFromCode(
-            transformedFile.code,
-            filename,
-          );
+          const script = this.createScriptFromCode(transformedCode, filename);
 
           const runScript = script.runInContext(
             vmContext,
@@ -851,7 +1012,7 @@ class Runtime {
         }
       }
     } else {
-      const script = this.createScriptFromCode(transformedFile.code, filename);
+      const script = this.createScriptFromCode(transformedCode, filename);
 
       const runScript = this._environment.runScript<RunScriptEvalResult>(
         script,
@@ -872,10 +1033,7 @@ class Runtime {
       return;
     }
 
-    const jestObject = this._createJestObjectFor(
-      filename,
-      localModule.require as LocalModuleRequire,
-    );
+    const jestObject = this._createJestObjectFor(filename);
 
     this.jestObjectCaches.set(filename, jestObject);
 
@@ -907,6 +1065,30 @@ class Runtime {
     this._currentlyExecutingModulePath = lastExecutingModulePath;
   }
 
+  private transformFile(
+    filename: string,
+    options?: InternalModuleOptions,
+  ): string {
+    const source = this.readFile(filename);
+
+    if (options?.isInternalModule) {
+      return source;
+    }
+
+    const transformedFile = this._scriptTransformer.transform(
+      filename,
+      this._getFullTransformationOptions(options),
+      source,
+    );
+
+    this._fileTransforms.set(filename, transformedFile);
+
+    if (transformedFile.sourceMapPath) {
+      this._sourceMapRegistry.set(filename, transformedFile.sourceMapPath);
+    }
+    return transformedFile.code;
+  }
+
   private createScriptFromCode(scriptSource: string, filename: string) {
     try {
       return new Script(this.wrapCodeInModuleWrapper(scriptSource), {
@@ -930,6 +1112,24 @@ class Runtime {
     }
 
     return require(moduleName);
+  }
+
+  private _importCoreModule(moduleName: string, context: VMContext) {
+    const required = this._requireCoreModule(moduleName);
+
+    return new SyntheticModule(
+      ['default', ...Object.keys(required)],
+      function () {
+        // @ts-ignore: TS doesn't know what `this` is
+        this.setExport('default', required);
+        Object.entries(required).forEach(([key, value]) => {
+          // @ts-ignore: TS doesn't know what `this` is
+          this.setExport(key, value);
+        });
+      },
+      // should identifier be `node://${moduleName}`?
+      {context, identifier: moduleName},
+    );
   }
 
   private _getMockedNativeModule(): typeof nativeModule.Module {
@@ -966,13 +1166,20 @@ class Runtime {
     // should we implement the class ourselves?
     class Module extends nativeModule.Module {}
 
+    Object.entries(nativeModule.Module).forEach(([key, value]) => {
+      // @ts-ignore
+      Module[key] = value;
+    });
+
     Module.Module = Module;
 
     if ('createRequire' in nativeModule) {
       Module.createRequire = createRequire;
     }
     if ('createRequireFromPath' in nativeModule) {
-      Module.createRequireFromPath = (filename: string | URL) => {
+      Module.createRequireFromPath = function createRequireFromPath(
+        filename: string | URL,
+      ) {
         if (typeof filename !== 'string') {
           const error = new TypeError(
             `The argument 'filename' must be string. Received '${filename}'.${
@@ -989,7 +1196,7 @@ class Runtime {
       };
     }
     if ('syncBuiltinESMExports' in nativeModule) {
-      Module.syncBuiltinESMExports = () => {};
+      Module.syncBuiltinESMExports = function syncBuiltinESMExports() {};
     }
 
     this._moduleImplementation = Module;
@@ -1001,12 +1208,14 @@ class Runtime {
     const modulePath =
       this._resolver.resolveStubModuleName(from, moduleName) ||
       this._resolveModule(from, moduleName);
-    if (!(modulePath in this._mockMetaDataCache)) {
+    if (!this._mockMetaDataCache.has(modulePath)) {
       // This allows us to handle circular dependencies while generating an
       // automock
 
-      this._mockMetaDataCache[modulePath] =
-        this._moduleMocker.getMetadata({}) || {};
+      this._mockMetaDataCache.set(
+        modulePath,
+        this._moduleMocker.getMetadata({}) || {},
+      );
 
       // In order to avoid it being possible for automocking to potentially
       // cause side-effects within the module environment, we need to execute
@@ -1030,36 +1239,39 @@ class Runtime {
             `See: https://jestjs.io/docs/manual-mocks.html#content`,
         );
       }
-      this._mockMetaDataCache[modulePath] = mockMetadata;
+      this._mockMetaDataCache.set(modulePath, mockMetadata);
     }
     return this._moduleMocker.generateFromMetadata(
-      this._mockMetaDataCache[modulePath],
+      // added above if missing
+      this._mockMetaDataCache.get(modulePath)!,
     );
   }
 
-  private _shouldMock(from: Config.Path, moduleName: string) {
+  private _shouldMock(from: Config.Path, moduleName: string): boolean {
     const explicitShouldMock = this._explicitShouldMock;
     const moduleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
       moduleName,
     );
     const key = from + path.delimiter + moduleID;
 
-    if (moduleID in explicitShouldMock) {
-      return explicitShouldMock[moduleID];
+    if (explicitShouldMock.has(moduleID)) {
+      // guaranteed by `has` above
+      return explicitShouldMock.get(moduleID)!;
     }
 
     if (
       !this._shouldAutoMock ||
       this._resolver.isCoreModule(moduleName) ||
-      this._shouldUnmockTransitiveDependenciesCache[key]
+      this._shouldUnmockTransitiveDependenciesCache.get(key)
     ) {
       return false;
     }
 
-    if (moduleID in this._shouldMockModuleCache) {
-      return this._shouldMockModuleCache[moduleID];
+    if (this._shouldMockModuleCache.has(moduleID)) {
+      // guaranteed by `has` above
+      return this._shouldMockModuleCache.get(moduleID)!;
     }
 
     let modulePath;
@@ -1068,35 +1280,35 @@ class Runtime {
     } catch (e) {
       const manualMock = this._resolver.getMockModule(from, moduleName);
       if (manualMock) {
-        this._shouldMockModuleCache[moduleID] = true;
+        this._shouldMockModuleCache.set(moduleID, true);
         return true;
       }
       throw e;
     }
 
     if (this._unmockList && this._unmockList.test(modulePath)) {
-      this._shouldMockModuleCache[moduleID] = false;
+      this._shouldMockModuleCache.set(moduleID, false);
       return false;
     }
 
     // transitive unmocking for package managers that store flat packages (npm3)
     const currentModuleID = this._resolver.getModuleID(
-      this._virtualMocks,
+      fromEntries(this._virtualMocks),
       from,
     );
     if (
-      this._transitiveShouldMock[currentModuleID] === false ||
+      this._transitiveShouldMock.get(currentModuleID) === false ||
       (from.includes(NODE_MODULES) &&
         modulePath.includes(NODE_MODULES) &&
         ((this._unmockList && this._unmockList.test(from)) ||
-          explicitShouldMock[currentModuleID] === false))
+          explicitShouldMock.get(currentModuleID) === false))
     ) {
-      this._transitiveShouldMock[moduleID] = false;
-      this._shouldUnmockTransitiveDependenciesCache[key] = true;
+      this._transitiveShouldMock.set(moduleID, false);
+      this._shouldUnmockTransitiveDependenciesCache.set(key, true);
       return false;
     }
-
-    return (this._shouldMockModuleCache[moduleID] = true);
+    this._shouldMockModuleCache.set(moduleID, true);
+    return true;
   }
 
   private _createRequireImplementation(
@@ -1109,18 +1321,37 @@ class Runtime {
     resolve.paths = (moduleName: string) =>
       this._requireResolvePaths(from.filename, moduleName);
 
-    const moduleRequire = (options && options.isInternalModule
+    const moduleRequire = (options?.isInternalModule
       ? (moduleName: string) =>
           this.requireInternalModule(from.filename, moduleName)
       : this.requireModuleOrMock.bind(
           this,
           from.filename,
         )) as LocalModuleRequire;
-    moduleRequire.cache = Object.create(null);
     moduleRequire.extensions = Object.create(null);
     moduleRequire.requireActual = this.requireActual.bind(this, from.filename);
     moduleRequire.requireMock = this.requireMock.bind(this, from.filename);
     moduleRequire.resolve = resolve;
+    moduleRequire.cache = (() => {
+      // TODO: consider warning somehow that this does nothing. We should support deletions, anyways
+      const notPermittedMethod = () => true;
+      return new Proxy<NodeJS.NodeRequireCache>(Object.create(null), {
+        defineProperty: notPermittedMethod,
+        deleteProperty: notPermittedMethod,
+        get: (_target, key) =>
+          typeof key === 'string' ? this._moduleRegistry.get(key) : undefined,
+        getOwnPropertyDescriptor() {
+          return {
+            configurable: true,
+            enumerable: true,
+          };
+        },
+        has: (_target, key) =>
+          typeof key === 'string' && this._moduleRegistry.has(key),
+        ownKeys: () => Array.from(this._moduleRegistry.keys()),
+        set: notPermittedMethod,
+      });
+    })();
 
     Object.defineProperty(moduleRequire, 'main', {
       enumerable: true,
@@ -1139,10 +1370,7 @@ class Runtime {
     return moduleRequire;
   }
 
-  private _createJestObjectFor(
-    from: Config.Path,
-    localRequire: LocalModuleRequire,
-  ): Jest {
+  private _createJestObjectFor(from: Config.Path): Jest {
     const disableAutomock = () => {
       this._shouldAutoMock = false;
       return jestObject;
@@ -1153,21 +1381,21 @@ class Runtime {
     };
     const unmock = (moduleName: string) => {
       const moduleID = this._resolver.getModuleID(
-        this._virtualMocks,
+        fromEntries(this._virtualMocks),
         from,
         moduleName,
       );
-      this._explicitShouldMock[moduleID] = false;
+      this._explicitShouldMock.set(moduleID, false);
       return jestObject;
     };
     const deepUnmock = (moduleName: string) => {
       const moduleID = this._resolver.getModuleID(
-        this._virtualMocks,
+        fromEntries(this._virtualMocks),
         from,
         moduleName,
       );
-      this._explicitShouldMock[moduleID] = false;
-      this._transitiveShouldMock[moduleID] = false;
+      this._explicitShouldMock.set(moduleID, false);
+      this._transitiveShouldMock.set(moduleID, false);
       return jestObject;
     };
     const mock: Jest['mock'] = (moduleName, mockFactory, options) => {
@@ -1176,11 +1404,11 @@ class Runtime {
       }
 
       const moduleID = this._resolver.getModuleID(
-        this._virtualMocks,
+        fromEntries(this._virtualMocks),
         from,
         moduleName,
       );
-      this._explicitShouldMock[moduleID] = true;
+      this._explicitShouldMock.set(moduleID, true);
       return jestObject;
     };
     const setMockFactory = (
@@ -1273,8 +1501,8 @@ class Runtime {
       isMockFunction: this._moduleMocker.isMockFunction,
       isolateModules,
       mock,
-      requireActual: localRequire.requireActual,
-      requireMock: localRequire.requireMock,
+      requireActual: this.requireActual.bind(this, from),
+      requireMock: this.requireMock.bind(this, from),
       resetAllMocks,
       resetModuleRegistry: resetModules,
       resetModules,
@@ -1355,11 +1583,52 @@ class Runtime {
     throw e;
   }
 
-  private getGlobalsForFile(from: Config.Path): JestGlobalsValues {
+  private getGlobalsForCjs(from: Config.Path): JestGlobalsValues {
     const jest = this.jestObjectCaches.get(from);
 
     invariant(jest, 'There should always be a Jest object already');
 
+    return {...this.getGlobalsFromEnvironment(), jest};
+  }
+
+  private async getGlobalsForEsm(
+    from: Config.Path,
+    context: VMContext,
+  ): Promise<VMModule> {
+    let jest = this.jestObjectCaches.get(from);
+
+    if (!jest) {
+      jest = this._createJestObjectFor(from);
+
+      this.jestObjectCaches.set(from, jest);
+    }
+
+    const globals: JestGlobalsValues = {
+      ...this.getGlobalsFromEnvironment(),
+      jest,
+    };
+
+    const module = new SyntheticModule(
+      Object.keys(globals),
+      function () {
+        Object.entries(globals).forEach(([key, value]) => {
+          // @ts-ignore: TS doesn't know what `this` is
+          this.setExport(key, value);
+        });
+      },
+      {context, identifier: '@jest/globals'},
+    );
+
+    await module.link(() => {
+      throw new Error('This should never happen');
+    });
+
+    await module.evaluate();
+
+    return module;
+  }
+
+  private getGlobalsFromEnvironment(): Omit<JestGlobalsValues, 'jest'> {
     return {
       afterAll: this._environment.global.afterAll,
       afterEach: this._environment.global.afterEach,
@@ -1370,12 +1639,23 @@ class Runtime {
       fdescribe: this._environment.global.fdescribe,
       fit: this._environment.global.fit,
       it: this._environment.global.it,
-      jest,
       test: this._environment.global.test,
       xdescribe: this._environment.global.xdescribe,
       xit: this._environment.global.xit,
       xtest: this._environment.global.xtest,
     };
+  }
+
+  private readFile(filename: Config.Path): string {
+    let source = this._cacheFS.get(filename);
+
+    if (!source) {
+      source = fs.readFileSync(filename, 'utf8');
+
+      this._cacheFS.set(filename, source);
+    }
+
+    return source;
   }
 }